web_app/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
email-validator>=2.1.1
